moved common things from core to different file, now only the job function is in os specific file
